requirements.txt
openai>=1.52.2
python-dotenv>=1.2.1
